Added capacity to block to prevent shrinking issue (see notes)
Previously, only the data length of a block was stored in the
block (header), and the capacity of the block was inferred by taking
the smallest power of two that would fit the data and the
header. However, this will be problematic if the data shrinks less
than half of its original capacity; this would cause the inferred
capacity to shrink, which would create a "gap" with no header after
the current shrunken block. To ensure that this doesn't happen (i.e.,
that the entire pool is a list of contiguous (but possibly deleted)
blocks, we need to keep track of when blocks are shrunken.

For the time being, no blocks are shrunken. This is only a very basic
allocation scheme emphasizing correctness over performance.